+++ src/node/create-babel-config.js
@@ -27,6 +27,7 @@ function createBabelConfig(
   const plugins = [
     require.resolve('babel-plugin-syntax-dynamic-import'),
     require.resolve('babel-plugin-transform-class-properties'),
+    require.resolve('babel-plugin-transform-object-rest-spread'),
     [
       require.resolve('@mapbox/babel-plugin-transform-jsxtreme-markdown'),
       {